refactor:extract date-range report layout to layouts/reports.py
